+++ integration_tests/suite/test_agents.py
@@ -5,6 +5,7 @@ from hamcrest import all_of, assert_that, calling, has_properties, is_, matches_
 
 from wazo_agentd_client.error import (
     AgentdClientError,
+    NOT_LOGGED,
     NO_SUCH_AGENT,
     NO_SUCH_LINE,
     UNAUTHORIZED,
@@ -90,7 +91,7 @@ class TestAgents(BaseIntegrationTest):
 
     @fixtures.user_line_extension(exten='1001', context='default', name_line='abcdef')
     @fixtures.agent(number='1234')
-    def test_login_logoff_user_agent_on_specific_line(self, user_line_extension, agent):
+    def test_login_pause_resume_logoff_user_agent_on_specific_line(self, user_line_extension, agent):
         self.create_user_token(user_line_extension['user_uuid'])
 
         with associations.user_agent(
@@ -106,6 +107,7 @@ class TestAgents(BaseIntegrationTest):
                     {
                         'id': agent['id'],
                         'logged': True,
+                        'paused': False,
                         'context': 'default',
                         'extension': '1001',
                         'number': '1234',
@@ -113,6 +115,18 @@ class TestAgents(BaseIntegrationTest):
                     }
                 ),
             )
+
+            # pause
+            self.agentd.agents.pause_user_agent()
+
+            status = self.agentd.agents.get_agent_status(agent['id'])
+            assert_that(status, has_properties(paused=True))
+
+            # resume
+            self.agentd.agents.resume_user_agent()
+
+            status = self.agentd.agents.get_agent_status(agent['id'])
+            assert_that(status, has_properties(paused=False))
 
             # logoff
             self.agentd.agents.logoff_user_agent()
@@ -124,6 +138,7 @@ class TestAgents(BaseIntegrationTest):
                     {
                         'id': agent['id'],
                         'logged': False,
+                        'paused': False,
                         'context': None,
                         'extension': None,
                         'number': '1234',
@@ -164,7 +179,7 @@ class TestAgents(BaseIntegrationTest):
             )
 
     @fixtures.user_line_extension(exten='1001', context='default', name_line='abcdef')
-    def test_login_logoff_user_agent_without_agent(self, user_line_extension):
+    def test_login_pause_resume_logoff_user_agent_without_agent(self, user_line_extension):
         self.create_user_token(user_line_extension['user_uuid'])
 
         # login
@@ -174,6 +189,19 @@ class TestAgents(BaseIntegrationTest):
             ),
             raises(AgentdClientError, has_properties(error=NO_SUCH_AGENT)),
         )
+
+        # pause
+        assert_that(
+            calling(self.agentd.agents.pause_user_agent),
+            raises(AgentdClientError, has_properties(error=NO_SUCH_AGENT)),
+        )
+
+        # resume
+        assert_that(
+            calling(self.agentd.agents.resume_user_agent),
+            raises(AgentdClientError, has_properties(error=NO_SUCH_AGENT)),
+        )
+
         # logoff
         assert_that(
             calling(self.agentd.agents.logoff_user_agent),
@@ -181,7 +209,7 @@ class TestAgents(BaseIntegrationTest):
         )
 
     @fixtures.user_line_extension(exten='1001', context='default', name_line='abcdef')
-    def test_login_logoff_user_agent_without_user(self, user_line_extension):
+    def test_login_pause_resume_logoff_user_agent_without_user(self, user_line_extension):
         self.create_user_token(UNKNOWN_UUID)
 
         # login
@@ -192,8 +220,57 @@ class TestAgents(BaseIntegrationTest):
             raises(AgentdClientError, has_properties(error=NO_SUCH_AGENT)),
         )
 
+        # pause
+        assert_that(
+            calling(self.agentd.agents.pause_user_agent),
+            raises(AgentdClientError, has_properties(error=NO_SUCH_AGENT)),
+        )
+
+        # resume
+        assert_that(
+            calling(self.agentd.agents.resume_user_agent),
+            raises(AgentdClientError, has_properties(error=NO_SUCH_AGENT)),
+        )
+
         # logoff
         assert_that(
             calling(self.agentd.agents.logoff_user_agent),
             raises(AgentdClientError, has_properties(error=NO_SUCH_AGENT)),
+
+        )
+
+    @fixtures.agent(id=42, number='1234')
+    @fixtures.user_line_extension(agentid=42, exten='1001', context='default', name_line='ab')
+    def test_pause_resume_logoff_user_agent_not_logged(self, agent, user_line_extension):
+        self.create_user_token(user_line_extension['user_uuid'])
+
+        # pause
+        assert_that(
+            calling(self.agentd.agents.pause_agent_by_number).with_args('1234'),
+            raises(AgentdClientError, has_properties(error=NOT_LOGGED)),
+        )
+        assert_that(
+            calling(self.agentd.agents.pause_user_agent),
+            raises(AgentdClientError, has_properties(error=NOT_LOGGED)),
+        )
+
+        # resume
+        assert_that(
+            calling(self.agentd.agents.unpause_agent_by_number).with_args('1234'),
+            raises(AgentdClientError, has_properties(error=NOT_LOGGED)),
+        )
+        assert_that(
+            calling(self.agentd.agents.resume_user_agent),
+            raises(AgentdClientError, has_properties(error=NOT_LOGGED)),
+        )
+
+        # logoff
+        assert_that(
+            calling(self.agentd.agents.logoff_agent_by_number).with_args('1234'),
+            raises(AgentdClientError, has_properties(error=NOT_LOGGED)),
+        )
+        assert_that(
+            calling(self.agentd.agents.logoff_user_agent),
+            raises(AgentdClientError, has_properties(error=NOT_LOGGED)),
+
         )