+++ wazo_agentd/http.py
@@ -267,6 +267,18 @@ class _PauseAgentByNumber(_BaseResource):
         return '', 204
 
 
+class _PauseUserAgent(_BaseResource):
+    @required_acl('agentd.users.me.agents.pause.create')
+    def post(self):
+        tenant_uuids = self._build_tenant_list({'recurse': True})
+        body = pause_schema.load(request.get_json())
+        user_uuid = token.user_uuid
+        self.service_proxy.pause_user_agent(
+            user_uuid, body['reason'], tenant_uuids=tenant_uuids
+        )
+        return '', 204
+
+
 class _UnpauseAgentByNumber(_BaseResource):
     @required_acl('agentd.agents.by-number.{agent_number}.unpause.create')
     def post(self, agent_number):
@@ -274,6 +286,15 @@ class _UnpauseAgentByNumber(_BaseResource):
         self.service_proxy.unpause_agent_by_number(
             agent_number, tenant_uuids=tenant_uuids
         )
+        return '', 204
+
+
+class _ResumeUserAgent(_BaseResource):
+    @required_acl('agentd.users.me.agents.resume.create')
+    def post(self):
+        tenant_uuids = self._build_tenant_list({'recurse': True})
+        user_uuid = token.user_uuid
+        self.service_proxy.resume_user_agent(user_uuid, tenant_uuids=tenant_uuids)
         return '', 204
 
 
@@ -291,6 +312,8 @@ class HTTPInterface:
         (_LogoffAgentById, '/agents/by-id/<int:agent_id>/logoff'),
         (_LogoffAgentByNumber, '/agents/by-number/<agent_number>/logoff'),
         (_LogoffUserAgent, '/users/me/agents/logoff'),
+        (_PauseUserAgent, '/users/me/agents/pause'),
+        (_ResumeUserAgent, '/users/me/agents/resume'),
         (_AddAgentToQueue, '/agents/by-id/<int:agent_id>/add'),
         (_RemoveAgentFromQueue, '/agents/by-id/<int:agent_id>/remove'),
         (_PauseAgentByNumber, '/agents/by-number/<agent_number>/pause'),